refactor: split modify rstest config hooks
